requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
pydantic
python-dotenv
requests
pandas
numpy
python-dateutil
pymysql
